Replaced the BET with an AST (again). The purpose for this is different than the origional AST which the BET replaced.
Here I'm introducing a new node type in order to hold pointers to scope frames. These form a separate tree, overlaid over the AST, pointing up to parent scopes.

The idea is that the tree of scopes is constructed during parse, but variable values are only populated within those at runtime, when AST nodes are executed.

The other point (which is important!) is that the scope nodes are freed during AST traversal, as their links point up to the root rather than down to leaves, and so cannot be traversed root-up.