The return of `CODE_ANON`. Again, this time for other reasons.
I found that using variables named `""` for anonymous blocks (and then invoking them in order to produce a single node for the code stack) doesn't work out, because in order to dereference that variable itself, we must introduce a new scope (otherwise one block's `""` would overwrite another's). But that scope means local variable assignment is not visible to the containing block, as intended. In other words, we can't just go introducing new scopes where the user doesn't expect.