Moved frame pointer to be common to all node types for code stacks.
I don't really like this, because it is unneccessary for many types to hold a frame pointer. However, various user-defined hooks in the future might make this neccessary. My goal for the moment is to provide a frame pointer for sake of `CODE_ANON`, and there are several other ways to achieve this. For example, a new code instruction, `CODE_FRAME` or somesuch, which sets the current frame during evaulation.

One alternate approach would be to keep a frame stack at runtime (instead of parse time). I am not doing this because I think that would mean blocks cannoy carry their parents lexical scope when called as commands.