Attached actions for pipe redirection lists to `#pipe`.
This is stunningly tricky; the parse order has `#exec` nodes produced first, but the pipes must be created beforehand during evaulation. That's especially awkward because there may be multiple `#exec` nodes for each pipe, and so a reference cannot be returned for `redir-list` to attach to. But the syntax for `redir-list` must occur after the left-side expression.

I solve this by pushing to a placeholder `&ci` list first, and then later appending that to the code being produced.